Move style functions and avoid being public
