+++ src/views/config_menu.rs
@@ -3,7 +3,7 @@ use iced::widget::Button;
 
 use crate::views::hardware_view::HardwareConnection;
 use crate::views::hardware_view::HardwareConnection::NoConnection;
-use crate::views::{menu_bar_button, menu_bar_highlight_button, menu_button};
+use crate::views::info_row::{menu_bar_button, menu_bar_highlight_button, menu_button};
 use iced::{Renderer, Theme};
 use iced_aw::menu::{Item, Menu};
