+++ src/views/devices_menu.rs
@@ -3,14 +3,12 @@ use crate::discovery::DiscoveredDevice;
 use crate::discovery::DiscoveryMethod::USBRaw;
 use crate::views::hardware_view::HardwareConnection;
 use crate::views::info_dialog::InfoDialogMessage::HardwareDetailsModal;
-use crate::views::info_row::{MENU_BUTTON_HOVER_STYLE, MENU_BUTTON_STYLE};
-use crate::views::menu_button;
+use crate::views::info_row::menu_button;
 #[cfg(feature = "usb")]
 use crate::views::ssid_dialog::SsidDialogMessage;
 use crate::HardwareConnection::*;
 use crate::Message;
 use iced::alignment;
-use iced::widget::button::Status::Hovered;
 use iced::widget::{button, text};
 use iced::widget::{horizontal_space, row};
 use iced::{Length, Renderer, Theme};
@@ -85,13 +83,7 @@ fn device_items<'a>(
                             hardware_details.clone(),
                             ssid_spec.as_ref().and_then(|_wf| ssid_spec.clone()),
                         )))
-                        .style(|_, status| {
-                            if status == Hovered {
-                                MENU_BUTTON_HOVER_STYLE
-                            } else {
-                                MENU_BUTTON_STYLE
-                            }
-                        }),
+                        .style(menu_button),
                 ));
             }
 
@@ -100,13 +92,7 @@ fn device_items<'a>(
                     button("Reset Device Wi-Fi to Default")
                         .width(Length::Fill)
                         .on_press(Message::ResetSsid(hardware_details.serial.clone()))
-                        .style(|_, status| {
-                            if status == Hovered {
-                                MENU_BUTTON_HOVER_STYLE
-                            } else {
-                                MENU_BUTTON_STYLE
-                            }
-                        }),
+                        .style(menu_button),
                 ));
             }
         }
